Move sampler and util to same module
